Update translation script
- Use `lupdate` from Qt6

The previous way on which the files were updated is done is by creating
some temporal files and then merging them to the original file.

Doing this merge would let us to NOT lost the translations already done.

It was done like that because we got the string from Python files and UI
files because `lupdate` from Qt5 couldn't extract strings from Python
files.

Now the `lupdate` from Qt6 can do, so we have no need to merge different
temp files, instead, we can do the update in a single command.

The new advantage is that the `<location>` tag is updated with all new
occurrences.